tweak intersection helpers to have fewer changes
