+++ src/helpers/intersection.ts
@@ -6,65 +6,53 @@
  * @return {{top: number, left: number, bottom: number, right: number}}
  */
 
-import {Point} from "../internalTypes";
+import {Point, AbsoluteRect, Rect} from "../internalTypes";
 
-export function getBoundingRectNoTransforms(el: HTMLElement): DOMRect {
+export function getBoundingRectNoTransforms(el: HTMLElement): Rect {
     let ta: string[];
 
     const rect = el.getBoundingClientRect();
     const style = getComputedStyle(el);
     const tx = style.transform;
 
-    if (!tx) {
-        return rect;
-    }
+    if (tx) {
+        let sx: number, sy: number, dx: number, dy: number;
+        if (tx.startsWith("matrix3d(")) {
+            ta = tx.slice(9, -1).split(/, /);
+            sx = +ta[0];
+            sy = +ta[5];
+            dx = +ta[12];
+            dy = +ta[13];
+        } else if (tx.startsWith("matrix(")) {
+            ta = tx.slice(7, -1).split(/, /);
+            sx = +ta[0];
+            sy = +ta[3];
+            dx = +ta[4];
+            dy = +ta[5];
+        } else {
+            return rect;
+        }
 
-    let sx: number, sy: number, dx: number, dy: number;
-    if (tx.startsWith("matrix3d(")) {
-        ta = tx.slice(9, -1).split(/, /);
-        sx = +ta[0];
-        sy = +ta[5];
-        dx = +ta[12];
-        dy = +ta[13];
-    } else if (tx.startsWith("matrix(")) {
-        ta = tx.slice(7, -1).split(/, /);
-        sx = +ta[0];
-        sy = +ta[3];
-        dx = +ta[4];
-        dy = +ta[5];
+        const to = style.transformOrigin;
+        const x = rect.x - dx - (1 - sx) * parseFloat(to);
+        const y = rect.y - dy - (1 - sy) * parseFloat(to.slice(to.indexOf(" ") + 1));
+        const w = sx ? rect.width / sx : el.offsetWidth;
+        const h = sy ? rect.height / sy : el.offsetHeight;
+
+        return {
+            x: x,
+            y: y,
+            width: w,
+            height: h,
+            top: y,
+            right: x + w,
+            bottom: y + h,
+            left: x
+        };
     } else {
         return rect;
     }
-
-    const to = style.transformOrigin;
-    const x = rect.x - dx - (1 - sx) * parseFloat(to);
-    const y = rect.y - dy - (1 - sy) * parseFloat(to.slice(to.indexOf(" ") + 1));
-    const w = sx ? rect.width / sx : el.offsetWidth;
-    const h = sy ? rect.height / sy : el.offsetHeight;
-
-    const newRect: Omit<DOMRect, "toJSON"> = {
-        x: x,
-        y: y,
-        width: w,
-        height: h,
-        top: y,
-        right: x + w,
-        bottom: y + h,
-        left: x
-    };
-
-    return {
-        ...newRect,
-        toJSON: () => JSON.stringify(newRect)
-    };
 }
-
-export type AbsoluteRect = {
-    top: number;
-    bottom: number;
-    left: number;
-    right: number;
-};
 
 /**
  * Gets the absolute bounding rect (accounts for the window's scroll position and removes transforms)
